feat(state-ownership): native-ize .new for Array/List/Hash/Map (ledger §D ③)
§D ③ ctor fork: drain the aggregate built-in constructors Array / List /
Positional / array and Hash / Map, which still bounced to the interpreter's
`dispatch_new`.

Unlike the pure-data static constructors, these are `&mut self` — shaped-
dimension parsing (`:shape(...)` + `:data`/positional population with the
X::Assignment shape errors), Slip/Range/Seq flattening, parameterized type
checks (`Array[Int].new`/`Hash[Int].new` -> `X::TypeCheck::Assignment`) and
container-metadata tagging. So, following the QuantHash pattern, the two arms
(~310 lines) are extracted wholesale into a new module
`methods_aggregate_ctor.rs` (`try_native_array_construct` /
`try_native_hash_construct`); the interpreter's `dispatch_new` arms now delegate
to them, and the VM fast path calls them via
`try_native_aggregate_construct_for_package` (resolves the parametric type name
into base + type args). True single impl, byte-identical between the two paths.

Measured: the `new` fallback for Array/List/Positional/array/Hash/Map drains to 0.

pin: t/native-aggregate-ctor.t (16, raku-parity subset). Shaped / Hash-named-data
mutsu-specific behaviour is covered by roast S02-types/array.t, hash.t and
S09-typed-arrays/*.
