+++ src/runtime/mod.rs
@@ -156,6 +156,7 @@ mod lock_reentry;
 mod main_args;
 mod metamodel;
 mod methods;
+mod methods_aggregate_ctor;
 mod methods_call_helpers;
 mod methods_classhow;
 mod methods_collection;
