+++ src/vm/vm_call_method_compiled.rs
@@ -175,6 +175,19 @@ impl Interpreter {
             self.method_dispatch_pure = true;
             return result;
         }
+        // Native aggregate construction: `Array`/`List`/`Positional`/`array`/
+        // `Hash`/`Map`.new(...) — shaped-dim parsing + parameterized type check +
+        // container-metadata tag, no env / registry / user code. Built directly via
+        // the single `try_native_array_construct` / `try_native_hash_construct`
+        // impls the interpreter's `dispatch_new` also delegates to.
+        if method == "new"
+            && let Value::Package(class_name) = &target
+            && let Some(result) =
+                self.try_native_aggregate_construct_for_package(*class_name, &args)
+        {
+            self.method_dispatch_pure = true;
+            return result;
+        }
         // Native built-in *class* method (a pure type-object method other than
         // `.new`, e.g. `Instant.from-posix`) — built directly instead of routing
         // through the interpreter's class-method dispatch.
@@ -1779,6 +1792,15 @@ impl Interpreter {
             && let Value::Package(class_name) = &target
             && let Some(result) =
                 self.try_native_quanthash_construct_for_package(*class_name, &args)
+        {
+            self.method_dispatch_pure = true;
+            return result;
+        }
+        // Native aggregate construction (mut path twin of the above).
+        if method == "new"
+            && let Value::Package(class_name) = &target
+            && let Some(result) =
+                self.try_native_aggregate_construct_for_package(*class_name, &args)
         {
             self.method_dispatch_pure = true;
             return result;
